Ok this has a lot of changes, I realized that my use of a bunch of vectors was not a great idea and wouldn't be easy to delete entities. so instead a map of ID, entity pairs is used to manage all entities. To keep track of different types sets containing just the IDs are used. To add items it works the same as previously, where any updatable are added to a queue and any item to be removed has its ID added to a set for destroyed entities. Before each update there is a preprocess that deletes all items that have their ID in the destroyed set and then adds the items from the queues.
Most of the changes present reflect this change in entity management, the rest are for properly handling destructors and making getters const.

What was called BaseEntity is now StaticEntity and a new BaseEntity class was added, this is so that any and all entities are derived from the same base class, which gives them an ID and message handler method. A virtual GetType method is added to the new BaseEntity, this is useful if something needs to know what type an Entity is i.e. missile batteries can set target using type, this gets all entities of that type and adds them to its targets vector.

Message struct and Message Dispatcher are added, these utilize the code from Matt Buckland's Game AI book, the main difference is that a message with a receiver ID < 0 will bet sent to all entities this is useful for clean up before destroying an entity since it may not know what has a pointer to it.